:fire: feat: Adding and removing movies

+++ server/src/domains/Movie/services/MovieService.ts
@@ -10,6 +10,7 @@ class MovieService {
      * @returns O filme criado.
      */
     async create(data: {
+        id: string;
         name: string;
         posterUrl: string;
         imdbGrade: number;
@@ -37,7 +38,7 @@ class MovieService {
      * @param id - ID do filme.
      * @returns O filme encontrado ou null se não existir.
      */
-    async findOne(id: number): Promise<Movie | null> {
+    async findOne(id: string): Promise<Movie | null> {
         return await prisma.movie.findUnique({
             where: { id },
         });
@@ -49,7 +50,7 @@ class MovieService {
      * @param data - Dados a serem atualizados.
      * @returns O filme atualizado.
      */
-    async update(id: number, data: Partial<Movie>): Promise<Movie> {
+    async update(id: string, data: Partial<Movie>): Promise<Movie> {
         return await prisma.movie.update({
             where: { id },
             data,
@@ -61,11 +62,27 @@ class MovieService {
      * @param id - ID do filme.
      * @returns O filme removido.
      */
-    async remove(id: number): Promise<Movie> {
+    async remove(id: string): Promise<Movie> {
+        // Verifique as dependências na tabela de reviews
+        const dependencies = await prisma.review.findMany({
+            where: { movieId: id },
+        });
+
+        if (dependencies.length > 0) {
+            throw new Error('Não é possível excluir o filme porque existem registros dependentes.');
+        }
+
+        // Remova todas as associações do filme com as bibliotecas antes de deletá-lo
+        await prisma.library_Movies.deleteMany({
+            where: { movieId: id },
+        });
+
+        // Agora, delete o filme da tabela de filmes
         return await prisma.movie.delete({
             where: { id },
         });
     }
+
 }
 
 export default new MovieService();
